Revise formula for start turn in Advanced Start
Now takes into account how many freebies the civs would've received without Advanced Start. [advc.250c]

Also refactored parts of change [advc.251] (start turn adjustment  on Immortal and Deity) and turned some statistics functions for floating point numbers [advc.003g] into function templates that also work with ScaledInt [advc.fract].

+++ CvGameCoreDLL/ScaledInt.h
@@ -205,7 +205,7 @@ public:
 	}
 	__forceinline ScaledInt(uint u) : m_i(static_cast<IntType>(SCALE * u))
 	{
-		FAssert(u <= INTMAX / SCALE);
+		FAssert(u <= static_cast<uint>(INTMAX) / static_cast<uint>(SCALE));
 	}
 	// Construction from rational
 	__forceinline ScaledInt(int iNum, int iDen)
